requirements.txt
argcomplete>=3.5.2
pyasn1~=0.6.1
pycryptodome~=3.23.0

